Add files via upload
This LLM Science Exam competition was held on Kaggle from Jul 11 to Oct 11, 2023. The goal was to help researchers better understand the ability of LLMs to test themselves, and the potential of LLMs that can be run in resource-constrained environments.
My score wasn't very high, but I joined the competition two weeks prior to the end.